replaced existing timer with new timer what fixes progress circle rendering problem on Android
